fix ci again
